Add `@noRd` label to documentation of R scripts
Add `@noRd` label to documentation of initial R scripts.  This
prevents the creation of `*.Rd` files and thus their inclusion in the
package index.  This is recommended for internal (i.e.,
non-user-facing) functions according to the following sources:

- https://community.rstudio.com/t/keywords-internal-vs-nord/35119
- https://style.tidyverse.org/documentation.html#internal-functions

At this time, the plan for the package is to provide one or a small
number of user-facing functions that utilizes more specialized (and
undocumented) functions such as those defined in the initial R
scripts.