Define alternative `outlier.rank()`
Define an alternative to `outlier.rank()`, tentatively named
`outlier.rank2()`.  It takes a vector of outlier statistics as input
rather than a matrix, along with a string giving the method with which
the outlier statistics were calculated: one of 'zrange.mean',
'zrange.median', 'zrange.trimmean', 'fraction.kmeans', or
'cosine.similarity'.  This is simpler and compatible with avoiding
unnecessary creation of matrices when separate vectors will suffice.
It returns a vector ranks rather than a matrix of ranks.