Precompute outlier stats in observed and null data
Add function `detect.outliers3()` to script `R/detect.outliers.R` that
precomputes the five outlier statistics for the observed and null
data.

Add function `calculate.p.values2()` to script
`R/calculate.p.values.R` that takes values of the five outlier
statistics for the observed transcript and vectors of the five outlier
statistics for the null data.  This obviates the need to run these
calculations in evaluating the initial outlier test, meaning that we
can skip directly to ranking the combined vector of the statistic in
the observed transcript and the statistics in the null transcripts,
for each of the five outlier statistics.  However, we will still need
to calculate the statistics for subsequent iterations, that is, if
there is at least one outlier in a given observed transcript.