Add `detect.outliers.precompute()`
Add draft function `detect.outliers.precompute()` to script
`R/detect.outliers.R`.  This alternate version of `detect.outliers()`
is intended to allow precomputing of outlier statistics for a
specified number (controlled by the new parameter
`num.null.precompute`) of iterations of the null data.
`detect.outliers3()` added precomputation of outlier statistics for
the complete null data set (as well as the complete observed data
set).  When an outlier is detected in an observed transcript, that
sample is excluded and computations repeated in the remaining
samples.  Because the identity of the most outlying sample will differ
from transcript to transcript, precomputing outlier statistics for the
observed data is not readily feasible.  However, with the null data,
the identity of the sample excluded after an outlier is detected is
arbitrary.  Previously a randomly chosen column (i.e., sample) from
the null data was excluded and then the outlier statistics
recalculated.  Assuming that this choice is truly irrelevant---which,
given that there are no outliers in the null data by construction,
seems reasonable---we could just as well drop the first column after
the first outlier is detected, the first and second columns after the
second outlier is detected, and so on.  Because the order of
exclusions is the same, we can precompute the outlier statistics on
the resulting subsets of the null data once before entering the loop
where p-values are calculated for each transcript.

This commit adds the new function, initializes a
list---`null.statistics.list`---to store the vectors of precomputed
outlier statistics for each iteration of the null data, and carries
out the computations.  A new version of `calculate.p.values()` will be
required that accepts this list of precomputed statistics, along with
logic to both use this new version of `calculate.p.values()` and also
to revert to the old version if an observed transcript contains more
outliers than we've precomputed.