Add to return value of `calculate.p.values()`
Add sample name, rank product, and (unadjusted) p-value to the outlier
test result return value of function `calculate.p.values()`.  In so
doing, change the data structure from a vector to a data
frame (because sample is a string, so we're mixing data types).  Also
change the name of that component of the return value from
'outlier.statistics.list' to 'results.list' and the name of each
element of this component from 'outlier.statistics.N' to 'round.N'.

Drop code to fill in missing p-values with `p.value.threshold`;
'missing p-values' are p-values associated with samples for whom the
outlier test was not run because we did not process the sample before
another sample that was found not to be an outlier.  This leaves them
as `NA` and also retains the exact (unadjusted) p-value of the last
processed sample.  Leaving them as `NA` reflects the fact that we did
not run the outlier test on these samples.