Create FDR matrix in `detect.outliers()`
Create a matrix of FDR-adjusted p-values (identical in structure to
the unadjusted p-values matrix) in `detect.outliers()`.  We build the
matrix through a somewhat-convoluted process of combining the data
frames of outlier test results and reshaping the resulting data frame
from long to wide format, then we fill in the matrix with the FDR
values.  As with the p-values matrix, entries in the FDR matrix
corresponding to samples and transcripts for which the outlier test
was not run (because the p-value associated with another sample
exceeded `p.value.threshold`) will be `NA`.

Also count the number of outliers based on the FDR-adjusted p-values
using a new parameter, `fdr.threshold`.  Change the existing outlier
counts vector to `num.outliers.unadjusted`, with the new counts vector
being `num.outliers.adjusted`.  Both are included in the return value
of `detect.outliers()`.

Update the documentation for `detect.outliers()` to reflect additions
and changes to the return value.