Add script to produce example data set
Add script `make.outliers.R` to produce example data set for vignette.
The script makes use of internal R objects `outliers.distributions` (a
vector) and `outliers.params` (a list) defined in
`R/sysdata.rda`; (these are made available to the developer via
`devtools::load_all()`).

The script is stored in `data-raw/` following the advice of "R
Packages" (2e) on keeping a record of how data sets were created:
https://r-pkgs.org/data.html#sec-data-data-raw.  It will not be
included with the bundled version of the package.